% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.r
\name{boxcox}
\alias{boxcox}
\title{Applies a Box-Cox transformation}
\usage{
boxcox(x, lambda)
}
\arguments{
\item{x}{The vector to be transformed}

\item{lambda}{The exponent}
}
\value{
A vector that is the Box-Cox transformation of \code{x}.
}
\description{
That's it -- this function just applies a Box-Cox transformation
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{boxcox} function is quite simple.
     \item There's nothing else to say about \code{boxcox} (assuming you know what a boxcox transformation is).
}
}
\examples{
boxcox(1:10,2)
boxcox(-5,0)
}
