% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.r
\name{pow}
\alias{pow}
\title{Raises a number to a power}
\usage{
pow(x, a = 2)
}
\arguments{
\item{x}{The vector to be powered}

\item{a}{The exponent}
}
\value{
A vector that is the power of \code{x}.
}
\description{
That's it -- this function just raises a number to a power
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{pow} function is quite simple.
     \item There's nothing else to say about \code{pow} (assuming you know what a boxcox transformation is).
}
}
\examples{
pow(1:10,2)
pow(-5,0)
}
